requirements.txt
Django==1.8
django-autoslug==1.9.3
python-decouple==3.1
pytz==2018.4
